Overlay added on homescreen and updated intro text
