--- tabs/src/App_bak.tsx
@@ -1,22 +1,0 @@
-import React from "react";
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import Home from "./Home";
-import Users from "./Users";
-import "./App.css";
-
-const App: React.FC = () => {
-  return (
-    <Router>
-      <div className="App">
-        <header className="App-header">
-          <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path="/users" element={<Users />} />
-          </Routes>
-        </header>
-      </div>
-    </Router>
-  );
-};
-
-export default App;
