+++ tabs/src/components/ErrorComponent.tsx
