+++ tabs/src/styles/Home.css
@@ -1,0 +1,27 @@
+.home-container {
+    position: relative;
+    width: 100%;
+    height: 95vh; /* Full viewport height */
+    background-image: url('../images/bg.png');
+    background-size: cover;
+    background-position: center;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.overlay {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5); /* Black transparent overlay */
+}
+
+.content {
+    position: relative;
+    z-index: 1; /* Ensure content is above the overlay */
+    color: white; /* Ensure text color is white */
+    text-align: center; /* Center text */
+}
